refactor: move View modules to modules dir
