+++ src/core-ts/modules/useInteractor.ts
