+++ src/core-ts/modules/useInteractorStyle.ts
