+++ src/core-ts/modules/useRenderWindowView.ts
